Move credentials provider mocks around the files (not operational)
